server/requirements.txt
Flask==0.10.1
graphql-core==0.4.7b1
graphql-epoxy==0.1a1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
six==1.10.0
Werkzeug==0.10.4
wheel==0.24.0
